+++ src/adc/api/__init__.py
@@ -2,5 +2,6 @@ from ._data import AudioData, FORMATS, FORMAT_MP3, FORMAT_WAV, FORMAT_EXTENSIONS
 from ._data import determine_audio_format_from_ext, determine_audio_format_from_bytes
 from ._data_types import DATATYPES, DATATYPE_CLASSIFICATION, DATATYPE_SPEECH, data_type_to_class
 from ._classification import AudioClassificationData
+from ._phonemizer import Phonemizer
 from ._speech import SpeechData
 from ._utils import locate_audio, load_audio_from_bytes, load_audio_from_file