+++ src/adc/phonemizer/__init__.py
@@ -1,0 +1,1 @@
+from ._passthrough import PassThrough